// src/app/services/api-backend/model/boleto.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Asiento } from './asiento';
import { Funcione } from './funcione';
import { TipoBoleto } from './tipoBoleto';
import { BoletosFactura } from './boletosFactura';


export interface Boleto { 
    codBoleto?: number;
    codFuncion?: number;
    codAsiento?: number;
    codTipoBoleto?: number;
    estado?: string | null;
    usuarioIng?: string | null;
    fechaIng?: string;
    usuarioAct?: string | null;
    fechaAct?: string | null;
    codAsientoNavigation?: Asiento;
    codFuncionNavigation?: Funcione;
    codTipoBoletoNavigation?: TipoBoleto;
    boletosFacturas?: Array<BoletosFactura> | null;
}