// src/app/services/api-backend/model/cliente.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Factura } from './factura';


export interface Cliente { 
    codCliente?: number;
    username?: string | null;
    password?: string | null;
    nombres?: string | null;
    apellidos?: string | null;
    telefono?: string | null;
    direccion?: string | null;
    nit?: string | null;
    estado?: string | null;
    correo?: string | null;
    usuarioIng?: string | null;
    fechaIng?: string;
    usuarioAct?: string | null;
    fechaAct?: string | null;
    facturas?: Array<Factura> | null;
}