// src/app/services/api-backend/model/factura.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MetodosPago } from './metodosPago';
import { Promociones } from './promociones';
import { BoletosFactura } from './boletosFactura';
import { Cliente } from './cliente';


export interface Factura { 
    codFactura?: number;
    codCliente?: number;
    codMetodoPago?: number;
    total?: number;
    iva?: number;
    estado?: string | null;
    idPromocion?: number | null;
    cantidadPromociones?: number | null;
    descuentoPromociones?: number | null;
    usuarioIng?: string | null;
    fechaIng?: string;
    usuarioAct?: string | null;
    fechaAct?: string | null;
    codClienteNavigation?: Cliente;
    codMetodoPagoNavigation?: MetodosPago;
    codPromocionNavigation?: Promociones;
    boletosFacturas?: Array<BoletosFactura> | null;
}