// src/app/services/api-backend/model/sala.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Sucursale } from './sucursale';
import { Asiento } from './asiento';
import { Funcione } from './funcione';
import { TiposSala } from './tiposSala';


export interface Sala { 
    codSala?: number;
    codSucursal?: number;
    codTipoSala?: number | null;
    nombre?: string | null;
    estado?: string | null;
    activa?: string | null;
    usuarioIng?: string | null;
    fechaIng?: string | null;
    usuarioAct?: string | null;
    fechaAct?: string | null;
    codSucursalNavigation?: Sucursale;
    codTipoSalaNavigation?: TiposSala;
    asientos?: Array<Asiento> | null;
    funciones?: Array<Funcione> | null;
}