// src/app/services/api-backend/model/todoCompra.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Factura } from './factura';
import { Boleto } from './boleto';


export interface TodoCompra { 
    boletos?: Array<Boleto> | null;
    factura?: Factura;
}